Moved test to `test_dependency_validation.py`
Test that app raises for improper config when a handler parameter is typed with a user-defined generic that is not marked with `Dependency` function, and the dependency is not provided.
